daemon/requirements.txt
websockets>=12.0
aiohttp>=3.9.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
